requirements.txt
streamlit>=1.35.0
pandas>=2.1.0
numpy>=1.26.0
yfinance>=0.2.40
ta>=0.11.0
scikit-learn>=1.4.0
httpx>=0.27.0
python-dotenv>=1.0.0
feedparser>=6.0.11
textblob>=0.18.0
apscheduler>=3.10.4
cachetools>=5.3.2
plotly>=5.20.0
lxml>=5.0.0
html5lib>=1.1
